Extract pagarme url to a settings file
