+++ pagarme/settings.py
@@ -1,0 +1,3 @@
+# encoding: utf-8
+
+BASE_URL = 'https://api.pagar.me/1/'
